+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- new `some example,tasks` command to run specific tasks
+
 ## [0.29.0] - 2019-04-09
 
 ### Added